+++ R/model_regression.R
@@ -494,9 +494,8 @@ specificationAdvice <- function(formula, data){
     if(length(suggests)>0){
       if(length(suggests) > 1)
         suggests <- paste0(suggests, collapse = ", ")
-      cat(paste("When testing for homophily,",
-                    "it is recommended to include all more fundamental effects.\n",
-                    "Try adding", suggests, "to the model specification.\n\n"))
+      snet_info("When testing for homophily, it is recommended to include all more fundamental effects.",
+                "Try adding {suggests} to the model specification.")
       }
   }
 }